break main into different files for seperate tasks
